requirements.txt
Django==1.11.11
django-appconf==1.0.2
django-compressor==2.2
django-crispy-forms==1.7.1
django-debug-toolbar==1.9.1
django-reversion==2.0.13
djangorestframework==3.7.7
mysqlclient==1.3.12
pkg-resources==0.0.0
python-memcached==1.59
pytz==2018.3
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
